refaktor RedigerAvansertBeregningSkjema til AvansertSkjema og sub-skjema(er)
