+++ src/components/AvansertSkjema/AvansertSkjemaForAndreBrukere/index.ts
@@ -1,0 +1,1 @@
+export { AvansertSkjemaForAndreBrukere } from './AvansertSkjemaForAndreBrukere'
