+++ src/pages/Beregning/BeregningAvansert.tsx
@@ -32,12 +32,10 @@ import {
   selectAarligInntektFoerUttakBeloep,
   selectIsEndring,
   selectLoependeVedtak,
-  selectNedreAldersgrense,
   selectEpsHarPensjon,
   selectEpsHarInntektOver2G,
   selectSivilstand,
 } from '@/state/userInput/selectors'
-import { getBrukerensAlderISluttenAvMaaneden } from '@/utils/alder'
 import { logger } from '@/utils/logging'
 
 import styles from './BeregningAvansert.module.scss'
@@ -56,7 +54,6 @@ export const BeregningAvansert: React.FC = () => {
   const aarligInntektFoerUttakBeloep = useAppSelector(
     selectAarligInntektFoerUttakBeloep
   )
-  const nedreAldersgrense = useAppSelector(selectNedreAldersgrense)
 
   const epsHarPensjon = useAppSelector(selectEpsHarPensjon)
   const epsHarInntektOver2G = useAppSelector(selectEpsHarInntektOver2G)
@@ -180,15 +177,7 @@ export const BeregningAvansert: React.FC = () => {
     <>
       {avansertSkjemaModus === 'redigering' && (
         <RedigerAvansertBeregning
-          gaaTilResultat={() => {
-            setAvansertSkjemaModus('resultat')
-            window.scrollTo(0, 0)
-          }}
           vilkaarsproeving={alderspensjon?.vilkaarsproeving}
-          brukerensAlderPlus1Maaned={getBrukerensAlderISluttenAvMaaneden(
-            person,
-            nedreAldersgrense
-          )}
         />
       )}
